requirements.txt
lark-parser==0.11.1
pytest==5.0.1
pytest-steps==1.7.3
pyverilog
sphinx==3.1.0
